# Improve performance and reduce allocations in StandardProvider