--- a/src/Ninject/Activation/Providers/StandardProvider.cs
+++ b/src/Ninject/Activation/Providers/StandardProvider.cs
@@ -84,16 +84,7 @@ namespace Ninject.Activation.Providers
             }
 
             var directive = this.DetermineConstructorInjectionDirective(context);
-
-            var arguments = directive.Targets.Select(target => this.GetValue(context, target)).ToArray();
-
-            var cachedInstance = context.Cache.TryGet(context);
-
-            if (cachedInstance != null)
-            {
-                return cachedInstance;
-            }
-
+            var arguments = GetValues(context, directive.Targets);
             return directive.Injector(arguments);
         }
 
@@ -108,10 +99,7 @@ namespace Ninject.Activation.Providers
             Ensure.ArgumentNotNull(context, "context");
             Ensure.ArgumentNotNull(target, "target");
 
-            var parameter = context
-                .Parameters.OfType<IConstructorArgument>()
-                .SingleOrDefault(p => p.AppliesToTarget(context, target));
-            return parameter != null ? parameter.GetValue(context, target) : target.ResolveWithin(context);
+            return GetValueCore(context, target);
         }
 
         /// <summary>
@@ -127,16 +115,53 @@ namespace Ninject.Activation.Providers
             return this.Type.ContainsGenericParameters ? this.Type.MakeGenericType(service.GetGenericArguments()) : this.Type;
         }
 
+        private static object GetValueCore(IContext context, ITarget target)
+        {
+            IConstructorArgument constructorArgument = null;
+
+            foreach (var parameter in context.Parameters)
+            {
+                if (parameter is IConstructorArgument ctorArg && ctorArg.AppliesToTarget(context, target))
+                {
+                    if (constructorArgument != null)
+                    {
+                        throw new InvalidOperationException("Sequence contains more than one matching element");
+                    }
+
+                    constructorArgument = ctorArg;
+                }
+            }
+
+            if (constructorArgument != null)
+            {
+                return constructorArgument.GetValue(context, target);
+            }
+
+            return target.ResolveWithin(context);
+        }
+
+        private static object[] GetValues(IContext context, ITarget[] targets)
+        {
+            object[] values = new object[targets.Length];
+
+            for (var i = 0; i < targets.Length; i++)
+            {
+                values[i] = GetValueCore(context, targets[i]);
+            }
+
+            return values;
+        }
+
         private ConstructorInjectionDirective DetermineConstructorInjectionDirective(IContext context)
         {
-            var directives = context.Plan.GetAll<ConstructorInjectionDirective>().ToList();
+            var directives = context.Plan.GetAll<ConstructorInjectionDirective>().ToArray();
 
-            if (directives.Count == 0)
+            if (directives.Length == 0)
             {
                 throw new ActivationException(ExceptionFormatter.NoConstructorsAvailable(context));
             }
 
-            if (directives.Count == 1)
+            if (directives.Length == 1)
             {
                 return directives[0];
             }
@@ -146,9 +171,9 @@ namespace Ninject.Activation.Providers
                     .GroupBy(directive => this.ConstructorScorer.Score(context, directive))
                     .OrderByDescending(g => g.Key)
                     .First()
-                    .ToList();
+                    .ToArray();
 
-            if (bestDirectives.Count > 1)
+            if (bestDirectives.Length > 1)
             {
                 throw new ActivationException(ExceptionFormatter.ConstructorsAmbiguous(context, bestDirectives));
             }