Improve performance and reduce allocations in StandardProvider: 
* Do not perform lookup in cache as this is already done right before we decide to create an instance.
* Reduce suboptimal usage of Linq
* Reduce allocations in case of default ctor. 
* Add benchmarks.

**Before:**

|                          Method |       Mean |     Error |    StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|-------------------------------- |-----------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
| Create_FromConstructorArguments | 3,680.1 ns | 27.251 ns | 25.491 ns |      0.5074 |           - |           - |              2136 B |
|             Create_FromBindings | 7,030.7 ns | 27.683 ns | 25.895 ns |      1.2512 |           - |           - |              5256 B |
|   Create_FromDefaultConstructor |   309.0 ns |  3.059 ns |  2.861 ns |      0.0815 |           - |           - |               344 B |

**After:**

|                          Method |       Mean |     Error |    StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|-------------------------------- |-----------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
| Create_FromConstructorArguments | 3,121.8 ns | 29.219 ns | 27.331 ns |      0.3777 |           - |           - |              1592 B |
|             Create_FromBindings | 6,683.2 ns | 81.384 ns | 76.127 ns |      1.1215 |           - |           - |              4712 B |
|   Create_FromDefaultConstructor |   260.7 ns |  3.183 ns |  2.977 ns |      0.0548 |           - |           - |               232 B |