@drieseng 2018-11-26
@scott-xu I know, but we're still targeting **.NET Framework 4.5** (and **.NET Standard 2.0**). We could work with **FEATURE** compile symbols to support "older" target frameworks, while still allowing us to consume new APIs.

This is what I did for **SSH.NET**. I (partially) documented our "switches" [here](https://github.com/sshnet/SSH.NET/wiki/Conditional-compilation-symbols).

Note that we should keep these to a minimum. For this aspect, **SSH.NET** is not the best example :-)